third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/nonces/scriptnonce-blocked-expected.txt

CONSOLE ERROR: Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js' because it violates the following Content Security Policy directive: "script-src 'nonce-nonce'". Note that 'script-src-elem' was not explicitly set, so 'script-src' is used as a fallback.

CONSOLE ERROR: Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js' because it violates the following Content Security Policy directive: "script-src 'nonce-notnonce'". Note that 'script-src-elem' was not explicitly set, so 'script-src' is used as a fallback.

Only the first two of these scripts should execute even though there are parse errors in the policy.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame3-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame4-->-->'
--------
PASS
